requirements.txt
requests>=2.31.0
chromadb==0.5.23
sentence-transformers==3.3.1
langchain==0.3.14
langchain-anthropic==0.3.3
langchain-chroma==0.1.4
anthropic==0.43.0
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
